requirements.txt
Flask==3.0.3
langchain==0.2.2
langchain_community==0.2.3
langchain_core==0.2.4
llama_index==0.10.43
marshmallow==3.21.3
pinecone==4.0.0
pydantic==2.7.3
python-dotenv==1.0.1
Requests==2.32.3
